Fixed error in typedef struct BPB in fat.h where the fat32 fields were incorrect. Fixed error in fat_detect_change() where it used incorrect values to detect the FAT type. Removed duplicate code from fat_find_internal().
Fixed error in fat_get_start_block() where it assumed the path contained a drive:
	* Added a drive parameter to fat_get_start_block()
	* Update calls to function in fat_find(), fat_create_int(), fat_create_subdirectory_entries(), fat_update_long_filename() to use drive parameter

Fixed error in fat_find() where it did not add the start of the data area in calls when using FAT32.
Fixed error in fat_find() where it did not update the findentry field in OPENFILE to skip over short filename entries for long filenames.