Replaced duplicate code to find data area relative to directory: 	* Moved code from fat_find(), fat_create_int(), fat_create_subdirectory_entries() to function get_block_data_area_relative_dir() 	* Inserted calls to get_block_data_area_relative_dir() in fat_find(), fat_create_int() 	* Removed datastart parameter from fat_create_entry(), fat_create_subdirectory_entries() and calls to fat_create_entry(), fat_create_subdirectory_entries()
Fixed error where it did not get the block number relative to directory when creating . and .. entries in a directory:
	* Added call to get_block_data_area_relative_dir() in fat_create_subdirectory_entries()
	* Updated call to blockio() to use new block number
	* Added bpb parameter to fat_create_subdirectory_entries()
	* Updated calls to fat_create_subdirectory_entries() in fat_create_entry(), fat_create_long_filename()
	* Added bpb parameter to fat_create_entry()
	* Updated calls to fat_entry_entry() in fat_create_int()

Fixed error in fat_create_entry() where it passed the directory name rather than the filename when creating a directory causing it to assume the start block was in the root directory.
Fixed error in fat_update_fat() where it assumed FAT size was in blocks not sectors.
Fixed error in fat_write() where it did not check if fat_update_fat() returned -1.
Fixed error in fat_update_fat() where a call to getblockdevice() checked if it returned NULL not -1.
Fixed error in fat_find_block() where it did not extract FAT 12 FAT entries correctly.